refactor: move some statics to MovedStatics
